+++ packages/ui-extensions/src/docs/shared/docs-type.ts
@@ -1,7 +1,0 @@
-import type {ReferenceEntityTemplateSchema} from '@shopify/generate-docs';
-
-export interface SharedReferenceEntityTemplateSchema
-  extends Pick<
-    ReferenceEntityTemplateSchema,
-    'name' | 'description' | 'category' | 'subCategory' | 'related'
-  > {}