+++ packages/ui-extensions-tester/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @shopify/ui-extensions-tester
 
+## 2026.7.0-rc.5
+
+### Patch Changes
+
+- Updated dependencies [[`be8c9e2`](https://github.com/Shopify/ui-extensions/commit/be8c9e206525a41f2544810a328d901e2cdcbb95)]:
+  - @shopify/ui-extensions@2026.7.0-rc.5
+
+## 2026.7.0-rc.4
+
+### Patch Changes
+
+- Updated dependencies [[`060e669`](https://github.com/Shopify/ui-extensions/commit/060e669fb5ff34b30f7e8a8fb1b2696e038a4938)]:
+  - @shopify/ui-extensions@2026.7.0-rc.4
+
 ## 2026.7.0-rc.3
 
 ### Patch Changes